+++ test/integration/fake_upstream.h
@@ -608,7 +608,7 @@ private:
     Network::ActiveUdpListenerFactory* udpListenerFactory() override {
       return udp_listener_factory_.get();
     }
-        envoy::api::v2::core::TrafficDirection direction() const override {
+    envoy::api::v2::core::TrafficDirection direction() const override {
       return envoy::api::v2::core::TrafficDirection::UNSPECIFIED;
     }