+++ test/integration/integration_admin_test.cc
@@ -122,6 +122,12 @@ std::string ContentType(const BufferingStreamDecoderPtr& response) {
 } // namespace
 
 TEST_P(IntegrationAdminTest, Admin) {
+
+  config_helper_.addConfigModifier([&](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
+    auto* outbound_listener = bootstrap.mutable_static_resources()->mutable_listeners(0);
+    outbound_listener->set_traffic_direction(envoy::api::v2::core::TrafficDirection::OUTBOUND);
+    outbound_listener->set_name("outbound_0");
+  });
   initialize();
 
   BufferingStreamDecoderPtr response = IntegrationUtil::makeSingleRequest(
@@ -414,6 +420,42 @@ TEST_P(IntegrationAdminTest, Admin) {
   envoy::admin::v2alpha::SecretsConfigDump secret_config_dump;
   config_dump.configs(5).UnpackTo(&secret_config_dump);
   EXPECT_EQ("secret_static_0", secret_config_dump.static_secrets(0).name());
+
+  // Validate that the inboundonly does not drain the outbound listeners.
+  response = IntegrationUtil::makeSingleRequest(lookupPort("admin"), "POST",
+                                                "/drain_listeners?inboundonly", "",
+                                                downstreamProtocol(), version_);
+  EXPECT_TRUE(response->complete());
+  EXPECT_EQ("200", response->headers().Status()->value().getStringView());
+  EXPECT_EQ("text/plain; charset=UTF-8", ContentType(response));
+  EXPECT_EQ(0, test_server_->counter("listener_manager.listener_removed")->value());
+
+  // Validate taht the outbound listener is drained on drain listeners.
+  response = IntegrationUtil::makeSingleRequest(lookupPort("admin"), "POST", "/drain_listeners", "",
+                                                downstreamProtocol(), version_);
+  EXPECT_TRUE(response->complete());
+  EXPECT_EQ("200", response->headers().Status()->value().getStringView());
+  EXPECT_EQ("text/plain; charset=UTF-8", ContentType(response));
+  test_server_->waitForCounterGe("listener_manager.listener_removed", 1);
+}
+
+// Validates that the inboundonly drains inbound listeners.
+TEST_P(IntegrationAdminTest, AdminDrainInboundOnly) {
+
+  config_helper_.addConfigModifier([&](envoy::config::bootstrap::v2::Bootstrap& bootstrap) -> void {
+    auto* inbound_listener = bootstrap.mutable_static_resources()->mutable_listeners(0);
+    inbound_listener->set_traffic_direction(envoy::api::v2::core::TrafficDirection::INBOUND);
+    inbound_listener->set_name("inbound_0");
+  });
+  initialize();
+
+  BufferingStreamDecoderPtr response = IntegrationUtil::makeSingleRequest(
+      lookupPort("admin"), "POST", "/drain_listeners?inboundonly", "", downstreamProtocol(),
+      version_);
+  EXPECT_TRUE(response->complete());
+  EXPECT_EQ("200", response->headers().Status()->value().getStringView());
+  EXPECT_EQ("text/plain; charset=UTF-8", ContentType(response));
+  test_server_->waitForCounterGe("listener_manager.listener_removed", 1);
 }
 
 TEST_P(IntegrationAdminTest, AdminOnDestroyCallbacks) {